Combining Zuite and DbZuite
The separation is a minuscule time saver since non db-tests avoid table truncation... In parctice, this is milliseconds shaved off an already speady test suite, and the code complexity doesn't make sense.
